+++ examples/simple/test/compound/simple/001-replay/_ansible/play_spec.rb
@@ -1,0 +1,7 @@
+describe 'ansible-playbook' do
+  subject { BarkingIguana::Compound::Ansible::ResultsParser.new(ENV['ANSIBLE_RESULTS_FILE']) }
+
+  it "did not change anything" do
+    expect(subject.recap.total_changes).to eq 0
+  end
+end