+++ examples/simple/test/compound/simple/001-replay/example-host/example_spec.rb
@@ -1,0 +1,5 @@
+describe file('/tmp/example-file') do
+  it "exists" do
+    expect(subject).to exist
+  end
+end