+++ resources/ansible_spec_helper.rb
@@ -1,0 +1,1 @@
+require 'barking_iguana/compound'